Удалил нативные sql запросы Переделал repository для успешной работы программы

+++ src/main/java/kg/Alessand/Task/controller/ParkController.java
@@ -38,8 +38,9 @@ public class ParkController {
         return parkService.comeInn(park);
     }
     @GetMapping("/findFreePlace")
-    private int findFreePlace(){
+    private long findFreePlace(){
         return parkService.findFreePlace();
     }
+
 
 }
